Refactor source properties into modular files
Split large SourcePropertiesPanel implementation into focused modules for clarity and testability. Added BibleSourceProperties.kt (Bible UI + loading helpers), CameraEnumeration.kt (camera/device/format enumeration and parsing), ScreenCaptureWindows.kt (per-OS window listing for screen capture), and SourcePropertyFields.kt (reusable property fields, sliders and FontDropdown). Cleaned SourcePropertiesPanel.kt by removing moved helpers and unused imports. This is a code-organization refactor with no intended behavioural changes; improves maintainability and enables easier testing.
